refactor: split live room standup components
